Updated poemwords and poemgame script. + monikaWin script.
Poemwords now has a selection of Monika's words + scores. Added ",0" on the end of all other words as a placeholder whilst we figure out how to score them for Monika (4th number is Monika's score per word). monikaWin script is a test script that executes if Monika is the winner of the poem game.